ecommerce/ecommerce/product/views.py
from django.shortcuts import render
from rest_framework import viewsets
from rest_framework.response import Response

from .models import Category
from .serializers import CategorySerializer


# Create your views here.
class CategoryViewSet(viewsets.ViewSet):
    """
    API endpoint that views all the categories
    """

    queryset = Category.objects.all()
    # serializer_class = CategorySerializer

    def list(self, request):
        serializer = CategorySerializer(self.queryset, many=True)
        return Response(serializer.data)
