ecommerce/requirements.txt
asgiref==3.8.1
astroid==3.1.0
black==24.4.0
click==8.1.7
dill==0.3.8
Django==5.0.4
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.1
exceptiongroup==1.2.0
flake8==7.0.0
flake8-for-pycharm==0.4.1
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
pycodestyle==2.11.1
pyflakes==3.2.0
pylint==3.1.0
pytest==8.1.1
pytest-django==4.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.11.0
